Created basic victory screen
Created a spectator pawn blueprint, arena gamemode blueprint, and victory screen widget blueprint

BP_ArenaGM tracks how many players are left in the game. If the player count reaches 1, it creates and displays the BP_VictoryScreen blueprint. It passes the name of the final player to display.

BP_VictoryScreen displays the name of the winning player and a text prompt to return to main menu